Update socket to send once.
