+++ src/main/java/com/rocketden/main/service/SubmitService.java
@@ -1,6 +1,5 @@
 package com.rocketden.main.service;
 
-import com.rocketden.main.dto.game.GameDto;
 import com.rocketden.main.dto.game.GameMapper;
 import com.rocketden.main.dto.game.SubmissionDto;
 import com.rocketden.main.dto.game.SubmissionRequest;
@@ -10,7 +9,6 @@ import com.rocketden.main.game_object.Player;
 import com.rocketden.main.game_object.PlayerCode;
 import com.rocketden.main.game_object.Submission;
 import com.rocketden.main.model.problem.Problem;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -20,13 +18,6 @@ import java.util.List;
  */
 @Service
 public class SubmitService {
-
-    private final SocketService socketService;
-
-    @Autowired
-    public SubmitService(SocketService socketService) {
-        this.socketService = socketService;
-    }
 
     // Test the submission and send a socket update.
     public SubmissionDto submitSolution(Game game, SubmissionRequest request) {
@@ -51,12 +42,26 @@ public class SubmitService {
         player.getSubmissions().add(submission);
         player.setSolved(true);
 
-        // Sort list of players by who is winning
-        GameDto gameDto = GameMapper.toDto(game);
-
-        // Send socket update with latest leaderboard info
-        socketService.sendSocketUpdate(gameDto);
-
         return GameMapper.submissionToDto(submission);
+    }
+
+    /**
+     * If all players have solved the problem, update the game as such.
+     * (Depending on the game setting, this may or may not end the game).
+     */
+    public void conditionalSolvedSocketMessage(Game game) {
+        // Variable to indicate whether all players have solved the problem.
+        boolean allSolved = true;
+        for (Player player : game.getPlayers().values()) {
+            if (player.getSolved() == null || !player.getSolved()) {
+                allSolved = false;
+                break;
+            }
+        }
+
+        // If the users have all completed the problem, end the game.
+        if (allSolved) {
+            game.setAllSolved(true);
+        }
     }
 }
